ArmPlatformPkg: PL061: fix accessing GPIO DATA
// All bits low except one bit high, restricted to 8 bits
// (i.e. ensures zeros above 8bits)

But '&&' is wrong at here. It'll only return 1 or 0 as the result
of GPIO_PIN_MASK_HIGH_8BIT(Pin).

Since PL061 spec said in below.

In order to write to GPIODATA, the corresponding bits in the mask,
resulting from the address bus, PADDR[9:2], must be HIGH. Otherwise
the bit values remain unchanged by the write.
Similarly, the values read from this register are determined for
each bit, by the mask bit derived from the address used to access
the data register, PADDR[9:2]. Bits that are 1 in the address mask
cause the corresponding bits in GPIODATA to be read, and bits that
are 0 in the address mask cause the corresponding bits in GPIODATA
to be read as 0, regardless of their value.

Now simplify the way to read/write bit of GPIO_DATA register.

Contributed-under: TianoCore Contribution Agreement 1.0
Signed-off-by: Haojian Zhuang <haojian.zhuang <at> linaro.org>